requirements.txt
opencv-python>=4.8.0

# Testing
pytest>=7.0
pyyaml>=6.0
